+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Grafana pod predicate to not trigger on pod deletion
 - Ensure organization is created before proceeding with datasources and sso settings
 
+### Fixes
+
+- Fix race condition when switching organization in Grafana client by using WithOrgID method
+
 ## [0.13.2] - 2025-02-06
 
 ### Added